requirements.txt
annotated-types==0.7.0
anyio==4.12.0
asttokens==3.0.1
attrs==25.4.0
beautifulsoup4==4.14.3
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.19
decorator==5.2.1
demjson3==3.0.6
executing==2.2.1
fonttools==4.61.1
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
gpt4-openai-api==0.9.0
grpcio==1.76.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
ipykernel==7.1.0
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.7.0
jupyter_core==5.9.1
kiwisolver==1.4.9
langchain==1.2.0
langchain-core==1.2.6
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.6.0
markdownify==1.2.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
numpy==2.4.0
orjson==3.11.5
ormsgpack==1.12.1
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.1.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
proto-plus==1.27.0
protobuf==4.25.1
psutil==7.2.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.1
PySocks==1.7.1
pyTelegramBotAPI==4.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.31.0
requests-toolbelt==1.0.0
rsa==4.9.1
seaborn==0.13.2
selenium==4.39.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.1
stack-data==0.6.3
tenacity==9.1.2
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
trio==0.32.0
trio-websocket==0.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
undetected-chromedriver==3.5.5
uritemplate==4.2.0
urllib3==2.6.2
uuid_utils==0.12.0
wcwidth==0.2.14
webdriver-manager==4.0.2
websocket-client==1.9.0
websockets==15.0.1
wsproto==1.3.2
xxhash==3.6.0
zstandard==0.25.0
